backend/requirements.txt
requests==2.31.0
selenium
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
openai
python-dotenv